requirements.txt
requests==2.28.1
python-dotenv==0.21.0
pyTelegramBotAPI==4.7.1
geopy==2.2.0
